feat: add enabled flag in order to activate/deactivate temporary files deletion

+++ tntconcept-binnacle/src/main/kotlin/com/autentia/tnt/binnacle/usecases/TemporaryAttachmentsDeletionUseCase.kt
@@ -1,5 +1,6 @@
 package com.autentia.tnt.binnacle.usecases
 
+import com.autentia.tnt.AppProperties
 import com.autentia.tnt.binnacle.core.domain.AttachmentInfo
 import com.autentia.tnt.binnacle.core.services.AttachmentFileSystemStorage
 import com.autentia.tnt.binnacle.repositories.AttachmentInfoRepository
@@ -11,17 +12,20 @@ import javax.transaction.Transactional
 class TemporaryAttachmentsDeletionUseCase internal constructor(
     private val attachmentInfoRepository: AttachmentInfoRepository,
     private val attachmentFileSystemStorage: AttachmentFileSystemStorage,
+    private val appProperties: AppProperties,
 ) {
 
     @Transactional
     fun delete() {
-        val temporaryAttachments = attachmentInfoRepository.findByIsTemporaryTrue()
+        if (appProperties.binnacle.temporaryAttachments.enabled) {
+            val temporaryAttachments = attachmentInfoRepository.findByIsTemporaryTrue()
 
-        val temporaryAttachmentsLessThanOneDay =
-            temporaryAttachments.filter { isMoreThanOneDay(it) }
+            val temporaryAttachmentsLessThanOneDay =
+                temporaryAttachments.filter { isMoreThanOneDay(it) }
 
-        temporaryAttachmentsLessThanOneDay.forEach { attachmentFileSystemStorage.deleteAttachmentFile(it.path) }
-        attachmentInfoRepository.delete(temporaryAttachmentsLessThanOneDay.map { it.id })
+            temporaryAttachmentsLessThanOneDay.forEach { attachmentFileSystemStorage.deleteAttachmentFile(it.path) }
+            attachmentInfoRepository.delete(temporaryAttachmentsLessThanOneDay.map { it.id })
+        }
     }
 
     fun isMoreThanOneDay(temporaryAttachment: AttachmentInfo) =
